remove unused properties and monitoring hack
