+++ webofneeds/conf/matcher-solr/matcher-solr.properties
@@ -38,22 +38,6 @@ matcher.solr.query.score.threshold=0.0
 # factor which is used to normalize the Solr score
 matcher.solr.query.score.normalizationFactor=0.001
 
-# if true, hints are generated and send to both needs (the incoming need and the matched need). if false, hints are
-# only generated for the incoming need (the matched needs do not receive any hints)
-matcher.solr.createHintsForBothNeeds=true
-
 # execute immediate commit of need documents in solr index after querying,
 # otherwise solr/solr configuration is responsible for commit
 matcher.solr.index.commit=false
-
-# turn this on if you want to monitor the system
-matcher.solr.monitoring=false
-
-# directory where monitoring statistics are periodically dumped (leave empty to let java create temp files)
-monitoring.output.dir=
-
-# milliseconds between two consecutive monitoring statistics dumps
-monitoring.interval.seconds=3600
-
-# reset the monitoring stats after output? (allows for comparing periods to track change more clearly)
-monitoring.reset.after.output=true
